requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==1.11
idna==2.8
Markdown==3.1.1
mistune==0.8.4
Pillow==6.1.0
pyasn1==0.4.6
Pygments==2.4.2
PyMySQL==0.9.3
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
ua-parser==0.8.0
urllib3==1.25.3
user-agents==2.0
